backend/requirements.txt
fastapi[standard]
python-magic
yara-python
uvicorn
pillow
cryptography